update login logout and header component
