+++ Tombolo/client-reactjs/src/components/layout/Header/ApplicationMenu.js
@@ -1,0 +1,44 @@
+import React from 'react';
+import { Dropdown, Space, Tooltip } from 'antd';
+import { AppstoreOutlined } from '@ant-design/icons';
+
+const ApplicationMenu = ({ applications, selected, handleApplicationChange }) => {
+  //build application list for dropdown
+  const getApplicationItems = (applications) => {
+    if (applications && applications.length > 0) {
+      return applications.map((app) => ({ key: app.value, label: app.display }));
+    } else {
+      return [];
+    }
+  };
+
+  return (
+    <Dropdown
+      menu={{
+        items: getApplicationItems(applications),
+        onClick: (e) => {
+          handleApplicationChange(e.key);
+        },
+      }}
+      placement="bottom"
+      trigger={['click']}>
+      <Tooltip title="Select an Application" placement="right">
+        <Space
+          style={{
+            color: 'white',
+            border: '1px solid white',
+            cursor: 'pointer',
+            padding: '0 4px',
+            borderRadius: '3px',
+            maxHeight: '32px',
+            minWidth: '200px',
+          }}>
+          <AppstoreOutlined />
+          <span> {selected}</span>
+        </Space>
+      </Tooltip>
+    </Dropdown>
+  );
+};
+
+export default ApplicationMenu;
